chore исправил парочку недочётов

+++ Assets/_Source/ScriptableObjects/Generators/GeneratorData.cs
@@ -5,11 +5,11 @@ public class GeneratorData : ScriptableObject
 {
     [SerializeField] private string _generatorName;
     [SerializeField] private double _baseCost;
-    [SerializeField] private float _costGrowthFactor;
+    [SerializeField] private double _costGrowthFactor;
     [SerializeField] private double _baseIncomePerSecond;
 
     public string GeneratorName => _generatorName;
     public double BaseCost => _baseCost;
-    public float CostGrowthFactor => _costGrowthFactor;
+    public double CostGrowthFactor => _costGrowthFactor;
     public double BaseIncomePerSecond => _baseIncomePerSecond;
 }
